Refactor tests for simplification
